--- iis-schemas/src/main/avro/eu/dnetlib/iis/referenceextraction/datasource/DocumentToDatasource.avdl
@@ -1,15 +1,0 @@
-@namespace("eu.dnetlib.iis.referenceextraction.datasource.schemas")
-protocol IIS{
-
-    record DocumentToDatasource {
-//      document identifier, foreign key: DocumentWithBasicMetadata.id ("document basic metadata" data store)
-        string documentId;
-//      identifier of the referenced datasource, 
-//      foreign key: Datasource.id
-        string datasourceId;
-//		Find more details on `confidenceLevel` constraints in eu/dnetlib/iis/README.markdown file.
-        float confidenceLevel;
-//      text snippet surrounding the matched reference, required mostly for internal debugging and analytics  
-        union { null , string } textsnippet = null;
-    }
-}
