+++ iis-schemas/src/main/avro/eu/dnetlib/iis/referenceextraction/service/DocumentToService.avdl
@@ -1,0 +1,15 @@
+@namespace("eu.dnetlib.iis.referenceextraction.service.schemas")
+protocol IIS{
+
+    record DocumentToService {
+//      document identifier, foreign key: DocumentWithBasicMetadata.id ("document basic metadata" data store)
+        string documentId;
+//      identifier of the referenced datasource, 
+//      foreign key: Service.id
+        string serviceId;
+//		Find more details on `confidenceLevel` constraints in eu/dnetlib/iis/README.markdown file.
+        float confidenceLevel;
+//      text snippet surrounding the matched reference, required mostly for internal debugging and analytics  
+        union { null , string } textsnippet = null;
+    }
+}
