+++ iis-wf/iis-wf-export-actionmanager/src/main/resources/eu/dnetlib/iis/wf/export/actionmanager/sequencefile/oozie_app/workflow.xml
@@ -14,7 +14,7 @@
             <value>$UNDEFINED$</value>
         </property>
         <property>
-            <name>input_document_to_datasource</name>
+            <name>input_document_to_service</name>
             <value>$UNDEFINED$</value>
         </property>
         <property>
@@ -58,7 +58,7 @@
             <description>Number of reducers to be used when exporting output. Affects number of output files.
                 When set to 0 reducing phase is omitted, input ordering is preserved and number of output files depends on number of mappers.
                 Used only for subset of inputs where large number of small files is provided, namely: 
-                input_document_to_project, input_document_to_dataset, input_document_to_datasource, input_document_to_document_classes, input_matched_doc_organizations.
+                input_document_to_project, input_document_to_dataset, input_document_to_service, input_document_to_document_classes, input_matched_doc_organizations.
                 For other inputs reducing phase is not executed.
             </description>
         </property>
@@ -291,6 +291,7 @@
             </main-class>
             <arg>eu.dnetlib.iis.referenceextraction.project.schemas.DocumentToProject</arg>
             <arg>eu.dnetlib.iis.referenceextraction.dataset.schemas.DocumentToDataSet</arg>
+            <arg>eu.dnetlib.iis.referenceextraction.service.schemas.DocumentToService</arg>
             <arg>eu.dnetlib.iis.export.schemas.DocumentToConceptIds</arg>
             <arg>eu.dnetlib.iis.documentsclassification.schemas.DocumentToDocumentClasses</arg>
             <arg>eu.dnetlib.iis.export.schemas.Citations</arg>
@@ -313,7 +314,7 @@
 
     <fork name="forking">
         <path start="decision-exporter-document-to-dataset" />
-        <path start="decision-exporter-document-to-datasource" />
+        <path start="decision-exporter-document-to-service" />
         <path start="decision-exporter-document-to-project" />
         <path start="decision-exporter-document-to-project-concepts" />
         <path start="decision-exporter-document-to-researchinitiatives" />
@@ -340,10 +341,10 @@
         </switch>
     </decision>
 
-	<decision name="decision-exporter-document-to-datasource">
+	<decision name="decision-exporter-document-to-service">
         <switch>
-            <case to="joining">${input_document_to_datasource eq "$UNDEFINED$"}</case>
-            <default to="exporter-document-to-datasource" />
+            <case to="joining">${input_document_to_service eq "$UNDEFINED$"}</case>
+            <default to="exporter-document-to-service" />
         </switch>
     </decision>
 
@@ -477,7 +478,7 @@
         <error to="fail" />
     </action>
 
-	<action name="exporter-document-to-datasource">
+	<action name="exporter-document-to-service">
         <map-reduce>
             <prepare>
                 <mkdir path="${nameNode}${output}/document_eoscServices" />
@@ -485,15 +486,15 @@
             <configuration>
                 <property>
                     <name>export.action.builder.factory.classname</name>
-                    <value>eu.dnetlib.iis.wf.export.actionmanager.module.DocumentToDatasourceActionBuilderModuleFactory</value>
+                    <value>eu.dnetlib.iis.wf.export.actionmanager.module.DocumentToServiceActionBuilderModuleFactory</value>
                 </property>
                 <property>
                     <name>avro.schema.input.key</name>
-                    <value>${wf:actionData('generate-schema')['eu.dnetlib.iis.referenceextraction.dataset.schemas.DocumentToDatasource']}</value>
+                    <value>${wf:actionData('generate-schema')['eu.dnetlib.iis.referenceextraction.service.schemas.DocumentToService']}</value>
                 </property>
                 <property>
                     <name>mapreduce.input.fileinputformat.inputdir</name>
-                    <value>${input_document_to_datasource}</value>
+                    <value>${input_document_to_service}</value>
                 </property>
                 <property>
                     <name>mapreduce.output.fileoutputformat.outputdir</name>
