--- iis-wf/iis-wf-export-actionmanager/src/test/resources/eu/dnetlib/iis/wf/export/actionmanager/sequencefile/sampledataproducer/input/document_to_datasource.json
@@ -1,1 +1,0 @@
-{"documentId":"id-10","datasourceId":"dsid-1","confidenceLevel":0.8164966, "textsnippet": null}
