+++ iis-wf/iis-wf-export-actionmanager/src/test/resources/eu/dnetlib/iis/wf/export/actionmanager/sequencefile/sampledataproducer/input/document_to_service.json
@@ -1,0 +1,1 @@
+{"documentId":"id-10","serviceId":"dsid-1","confidenceLevel":0.8164966, "textsnippet": null}
