--- iis-wf/iis-wf-primary/src/test/resources/eu/dnetlib/iis/wf/primary/processing/sampledataproducer/output/document_to_eoscservice.json
@@ -1,6 +1,0 @@
-{
-  "documentId": "id-30",
-  "datasourceId": "10|eosc________::c491811e9a6afa69cdcab0f92fca6f7b",
-  "confidenceLevel": 0.8,
-  "textsnippet": "extraction algorithm.  Madis is heavily used in IIS:  https://github.com/openaire/iis "
-}
