+++ iis-wf/iis-wf-primary/src/test/resources/eu/dnetlib/iis/wf/primary/processing/sampledataproducer/output/document_to_service.json
@@ -1,0 +1,6 @@
+{
+  "documentId": "id-30",
+  "serviceId": "10|eosc________::c491811e9a6afa69cdcab0f92fca6f7b",
+  "confidenceLevel": 0.8,
+  "textsnippet": "extraction algorithm.  Madis is heavily used in IIS:  https://github.com/openaire/iis "
+}
