--- iis-wf/iis-wf-primary/src/test/resources/eu/dnetlib/iis/wf/primary/processing/sampledataproducer/output/report/document_to_eoscservice.json
@@ -1,1 +1,0 @@
-{"key":"processing.referenceExtraction.eoscservice.references", "type":"COUNTER", "value": "1"}
