+++ iis-wf/iis-wf-primary/src/test/resources/eu/dnetlib/iis/wf/primary/processing/sampledataproducer/output/report/document_to_service.json
@@ -1,0 +1,1 @@
+{"key":"processing.referenceExtraction.service.references", "type":"COUNTER", "value": "1"}
