+++ iis-wf/iis-wf-primary/src/test/resources/eu/dnetlib/iis/wf/primary/processing/sampledataproducer/output/report/primary-processing-execution-times.json
@@ -9,7 +9,7 @@
 {"key": "processing.referenceExtraction.researchInitiative.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 780000])"}
 {"key": "processing.referenceExtraction.community.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 500000])"}
 {"key": "processing.referenceExtraction.covid-19.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 500000])"}
-{"key": "processing.referenceExtraction.eoscservice.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 500000])"}
+{"key": "processing.referenceExtraction.service.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 500000])"}
 {"key": "processing.referenceExtraction.softwareUrl.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 700000])"}
 {"key": "processing.referenceExtraction.patent.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 700000])"}
 {"key": "processing.docOrgMatching.affiliationBased.duration", "type": "DURATION", "value": "$(LONG_RANGE: [1000, 200000])"}
