+++ iis-wf/iis-wf-referenceextraction/src/main/java/eu/dnetlib/iis/wf/referenceextraction/service/ServiceDBBuilder.java
@@ -1,0 +1,46 @@
+package eu.dnetlib.iis.wf.referenceextraction.service;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Map;
+
+import org.apache.commons.lang.StringUtils;
+
+import com.google.common.base.Preconditions;
+
+import eu.dnetlib.iis.importer.schemas.Service;
+import eu.dnetlib.iis.wf.referenceextraction.AbstractDBBuilder;
+
+/**
+ * Process creating a database reading {@link Service} input avro records.
+ *
+ * @author mhorst
+ */
+public class ServiceDBBuilder extends AbstractDBBuilder<Service> {
+    
+    private final static String PARAM_SCRIPT_LOCATION = "scriptLocation";
+
+    // -------------------------- CONSTRUCTORS ------------------------------
+
+    public ServiceDBBuilder() {
+        super(Service.SCHEMA$, "service", "service_db");
+    }
+
+    // -------------------------- LOGIC -------------------------------------
+
+    @Override
+    public ProcessExecutionContext initializeProcess(Map<String, String> parameters) throws IOException {
+        String scriptLocation = parameters.get(PARAM_SCRIPT_LOCATION);
+        Preconditions.checkArgument(StringUtils.isNotBlank(scriptLocation),
+                "sql script location not provided, '%s' parameter is missing!", PARAM_SCRIPT_LOCATION);
+
+        String targetDbLocation = System.getProperty("java.io.tmpdir") + File.separatorChar + "services.db";
+        File targetDbFile = new File(targetDbLocation);
+        targetDbFile.setWritable(true);
+
+        return new ProcessExecutionContext(
+                Runtime.getRuntime().exec("python scripts/madis/mexec.py -w " + targetDbLocation + " -f " + scriptLocation),
+                targetDbFile);
+        
+    }
+}
