--- iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/main/oozie_app/import.txt
@@ -1,3 +1,0 @@
-## This is a classpath-based import file (this header is required)
-sqlite_builder classpath eu/dnetlib/iis/wf/referenceextraction/eoscservice/sqlite_builder/oozie_app
-main_sqlite classpath eu/dnetlib/iis/wf/referenceextraction/eoscservice/main_sqlite/oozie_app
