--- iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/main_sqlite/oozie_app/lib/scripts/eosc.sql
@@ -1,5 +1,0 @@
-create temp table pubs as setschema 'c1,c2' select jsonpath(c1, '$.id', '$.text') from stdinput();
-
-select jdict('documentId', id, 'datasourceId', sid, 'confidenceLevel', 0.8, 'textsnippet', j2s(prev,middle,next)) from 
-(setschema 'id,prev,middle,next' select c1 as id,textwindow2s(regexpr("\n",c2," "),10,1,1,"https*://.*") from pubs), services 
-where lower(regexpr("https*://(www\d*\.)*",stripchars(regexpr("(https*://.*)",middle),"\/,.;_()"),"")) =  url and not regexprmatches(".*/.*",next) group by id,sid;
