--- iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/sqlite_builder/oozie_app/workflow.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0"?>
-<workflow-app xmlns="uri:oozie:workflow:0.4" name="referenceextraction_eosc_services_sqlite_builder">
-	
-	<parameters>
-		<property>
-			<name>input_service</name>
-			<description>input services datastore with eu.dnetlib.iis.importer.schemas.Service avro records</description>
-		</property>
-		<property>
-			<name>output_service_db</name>
-			<description>output services SQLite DB path</description>
-		</property>
-	</parameters>
-
-    <global>
-        <job-tracker>${jobTracker}</job-tracker>
-        <name-node>${nameNode}</name-node>
-        <configuration>
-            <property>
-                <name>mapreduce.job.queuename</name>
-                <value>${queueName}</value>
-            </property>
-            <property>
-                <name>oozie.launcher.mapred.job.queue.name</name>
-                <value>${oozieLauncherQueueName}</value>
-            </property>
-        </configuration>
-    </global>
-    
-    
-	<start to="sqlite_builder" />
-
-    <action name="sqlite_builder">
-        <java>
-            <!-- This is simple wrapper for the Java code -->
-            <main-class>eu.dnetlib.iis.common.java.ProcessWrapper</main-class>
-            <!-- The business Java code that gets to be executed -->
-            <arg>eu.dnetlib.iis.wf.referenceextraction.eoscservices.ServiceDBBuilder</arg>
-            <!-- All input and output ports have to be bound to paths in 
-            HDFS, working directory has to be specified as well -->
-            <arg>-Iservice=${input_service}</arg>
-            <arg>-Oservice_db=${output_service_db}</arg>
-            <arg>-PscriptLocation=scripts/buildeoscdb.sql</arg>
-        </java>
-        <ok to="end"/>
-        <error to="fail"/>
-    </action>
-    
-    <kill name="fail">
-        <message>Unfortunately, the process failed -- error message:
-        			[${wf:errorMessage(wf:lastErrorNode())}]
-        		</message>
-    </kill>
-
-    <end name="end"/>
-</workflow-app>
