+++ iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/service/main/oozie_app/import.txt
@@ -1,0 +1,3 @@
+## This is a classpath-based import file (this header is required)
+sqlite_builder classpath eu/dnetlib/iis/wf/referenceextraction/service/sqlite_builder/oozie_app
+main_sqlite classpath eu/dnetlib/iis/wf/referenceextraction/service/main_sqlite/oozie_app
