+++ iis-wf/iis-wf-referenceextraction/src/main/resources/eu/dnetlib/iis/wf/referenceextraction/service/sqlite_builder/oozie_app/lib/scripts/buildeoscdb.sql
