--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/data/document_to_service.json
@@ -1,1 +1,0 @@
-{"documentId": "50|od_______307::c3297980394e4f10e5bc6b038f1a9095", "datasourceId": "10|eosc________::c8c6e8d211d6df4ee8a187fa1134bd92", "confidenceLevel": 0.8, "textsnippet": "The PROVIDE dashboard (https://provide.openaire.eu) target"}
