--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/data/report_no_matches.json
@@ -1,1 +1,0 @@
-{"key":"processing.referenceExtraction.eoscservice.references", "type":"COUNTER", "value": "0"}
