--- iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/eoscservice/main/sampletest_empty_text_input/oozie_app/import.txt
@@ -1,2 +1,0 @@
-## This is a classpath-based import file (this header is required)
-referenceextraction_eoscservice classpath eu/dnetlib/iis/wf/referenceextraction/eoscservice/main/oozie_app
