+++ iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/service/data/report.json
@@ -1,0 +1,1 @@
+{"key":"processing.referenceExtraction.service.references", "type":"COUNTER", "value": "1"}
