+++ iis-wf/iis-wf-referenceextraction/src/test/resources/eu/dnetlib/iis/wf/referenceextraction/service/main/sampletest_empty_text_input/oozie_app/import.txt
@@ -1,0 +1,2 @@
+## This is a classpath-based import file (this header is required)
+referenceextraction_service classpath eu/dnetlib/iis/wf/referenceextraction/service/main/oozie_app
